+++ lua/pomo/init.lua
@@ -17,10 +17,12 @@ end
 ---@param time_limit integer The time limit, in seconds.
 ---@param name string|?
 ---@param repeat_n integer|? The number of the times to repeat the timer.
+---@param cfg pomo.Config|? Override the config.
 ---@return integer time_id
-M.start_timer = function(time_limit, name, repeat_n)
+M.start_timer = function(time_limit, name, repeat_n, cfg)
+  cfg = cfg and cfg or M.get_config()
   local timer_id = timers:first_available_id()
-  local timer = Timer.new(timer_id, time_limit, name, M.get_config(), repeat_n)
+  local timer = Timer.new(timer_id, time_limit, name, cfg, repeat_n)
 
   timers:store(timer)